Restructured Common folder to the root
