+++ lib/config/routing/router.dart
@@ -1,9 +1,9 @@
-import 'package:flex_workout_logger/features/common/ui/screens/error_screen.dart';
-import 'package:flex_workout_logger/features/common/ui/utils/ui_extensions.dart';
+import 'package:flex_workout_logger/config/routing/error_screen.dart';
 import 'package:flex_workout_logger/features/exercises/ui/screens/exercises_create_screen.dart';
 import 'package:flex_workout_logger/features/exercises/ui/screens/exercises_edit_screen.dart';
 import 'package:flex_workout_logger/features/exercises/ui/screens/exercises_view_screen.dart';
 import 'package:flex_workout_logger/features/exercises/ui/screens/library_screen.dart';
+import 'package:flex_workout_logger/utils/ui_extensions.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
