--- lib/features/common/infrastructure/schemas/test_realm_model.dart
@@ -1,13 +1,0 @@
-import 'package:realm/realm.dart';
-
-part 'test_realm_model.g.dart';
-
-@RealmModel()
-class _Car {
-  @PrimaryKey()
-  late ObjectId id;
-
-  late String make;
-  late String? model;
-  late int? miles;
-}
