+++ lib/features/exercises/controllers/exercises_list_controller.dart
@@ -1,6 +1,5 @@
 import 'package:flex_workout_logger/features/exercises/domain/entities/exercise_entity.dart';
 import 'package:flex_workout_logger/features/exercises/providers.dart';
-import 'package:fpdart/fpdart.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'exercises_list_controller.g.dart';
@@ -45,7 +44,7 @@ class ExercisesListController extends _$ExercisesListController {
     final items = state.valueOrNull ?? [];
 
     state = const AsyncValue.loading();
-    
+
     items.remove(entity);
     state = AsyncValue.data(items);
   }
