+++ lib/features/exercises/ui/screens/exercises_create_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flex_workout_logger/config/theme/app_layout.dart';
-import 'package:flex_workout_logger/features/common/ui/utils/ui_extensions.dart';
-import 'package:flex_workout_logger/features/common/ui/widgets/step_indicator.dart';
 import 'package:flex_workout_logger/features/exercises/ui/widgets/exercise_create_form.dart';
+import 'package:flex_workout_logger/utils/ui_extensions.dart';
+import 'package:flex_workout_logger/widgets/step_indicator.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
