+++ lib/features/exercises/ui/screens/library_screen.dart
@@ -1,5 +1,5 @@
-import 'package:flex_workout_logger/features/common/ui/utils/ui_extensions.dart';
 import 'package:flex_workout_logger/features/exercises/ui/widgets/exercises_list.dart';
+import 'package:flex_workout_logger/utils/ui_extensions.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
