+++ test/mocks/test_exercises.dart
@@ -1,6 +1,6 @@
 import 'package:faker/faker.dart';
-import 'package:flex_workout_logger/features/common/ui/utils/date_time_extensions.dart';
 import 'package:flex_workout_logger/features/exercises/infrastructure/schema.dart';
+import 'package:flex_workout_logger/utils/date_time_extensions.dart';
 import 'package:realm/realm.dart';
 
 class MockExercises {
